splitting off natlinkconfig_cli.py from natlinkconfigfunctions.py got rid of wxdialogs.py, in favor of tkinter_dialogs.py
